Refactor negamax with enum Actual/Pruned return
