+++ src/solver/game_table/table.rs
@@ -1,10 +1,12 @@
+use super::NegamaxResult;
+
 pub(super) struct Table {
-  t: Vec<Option<(u64, i8)>>
+  t: Vec<Option<(u64, NegamaxResult)>>
 }
 
 impl Table {
   pub fn new(size: usize) -> Self {
-    let t: Vec<Option<(u64, i8)>> = vec![None; size];
+    let t: Vec<Option<(u64, NegamaxResult)>> = vec![None; size];
     Self {
       t
     }
@@ -20,18 +22,18 @@ impl Table {
     k as usize % self.t.len()
   }
 
-  pub fn put(&mut self, k: u64, v: i8) {
+  pub fn put(&mut self, k: u64, v: NegamaxResult) {
     let i = self.index(k);
     self.t[i] = Some((k, v));
   }
 
-  pub fn get(&self, k: u64) -> Option<i8> {
+  pub fn get(&self, k: u64) -> Option<&NegamaxResult> {
     let i = self.index(k);
     let curr = &self.t[i];
     
     if let Some((k0, v0)) = curr {
       if *k0 == k {
-        Some(*v0)
+        Some(v0)
       } else {
         None
       }
